+++ src/main/java/edu/isep/easypark/DAOimpl.java
@@ -7,6 +7,13 @@ import javax.sql.DataSource;
 import org.springframework.jdbc.core.BeanPropertyRowMapper;
 import org.springframework.jdbc.core.JdbcTemplate;
 
+import edu.isep.easypark.model.Comment;
+import edu.isep.easypark.model.Commentaires;
+import edu.isep.easypark.model.ParkingLot;
+import edu.isep.easypark.model.User;
+import edu.isep.easypark.model.UserInformations;
+import edu.isep.easypark.model.Voiture;
+
 public class DAOimpl {
 	private JdbcTemplate jdbcTemplate;
 
@@ -40,11 +47,11 @@ public class DAOimpl {
 	
 	public List<ParkingLot> fillParkingLot2(ParkingLot parkingLot) {
 	
-	parkingLot.adresse="%".concat(parkingLot.adresse).concat("%"); //SQL à l'air d'échapper les %
+	parkingLot.setAdresse("%".concat(parkingLot.getAdresse()).concat("%")); //SQL à l'air d'échapper les %
 	     String sql = "select * from place where adresse like ? and id_user like ?";
 
 	           return this.jdbcTemplate.query(sql, new BeanPropertyRowMapper<ParkingLot>(
-						ParkingLot.class),  parkingLot.adresse, parkingLot.id_user);
+						ParkingLot.class),  parkingLot.getAdresse(), parkingLot.getId_user());
 	 
 	                      }
 	
@@ -53,8 +60,8 @@ public class DAOimpl {
 	public int insertPlace(ParkingLot place){
 			
 	String query = "insert into place (nom, ville, adresse, superficie, horaires, type, commentaire, id_user,note) values(?, ?, ?, ?, ?, ?, ?,?,?)";
-		int statut = this.jdbcTemplate.update(query, new Object[] { place.nom, place.ville,
-				place.adresse, place.superficie, place.horaires, place.type, place.commentaire, place.id_user,place.note});
+		int statut = this.jdbcTemplate.update(query, new Object[] { place.getNom(), place.getVille(),
+				place.getAdresse(), place.getSuperficie(), place.getHoraires(), place.getType(), place.getCommentaire(), place.getId_user(),place.getNote()});
 			
 		return statut;
 	
@@ -65,10 +72,10 @@ public class DAOimpl {
 		String query = "insert into user (email,password) values(?, ?)";
 		String query2 = "insert into user_informations (id_user, name, firstname) values (?,?,?)";
 		
-		this.jdbcTemplate.update(query, new Object[] { user.email, user.password});
+		this.jdbcTemplate.update(query, new Object[] { user.getEmail(), user.getPassword()});
 		int id_user = getUserId(user);
 		
-		this.jdbcTemplate.update(query2, new Object[] { id_user, userInformations.name, userInformations.firstname});
+		this.jdbcTemplate.update(query2, new Object[] { id_user, userInformations.getName(), userInformations.getFirstname()});
 
 		
 		return id_user;
@@ -88,7 +95,7 @@ public class DAOimpl {
 		boolean result = false;
 	 
 		int count = this.jdbcTemplate.queryForObject(
-	                        sql, new Object[] { user.email, user.password }, Integer.class);
+	                        sql, new Object[] { user.getEmail(), user.getPassword() }, Integer.class);
 
 
 		if (count > 0) {
@@ -101,7 +108,7 @@ public class DAOimpl {
 	 public int getUserId(User user){
 		 String sql = "SELECT ID FROM user WHERE email = ? AND password = ?";
 		 int id =  this.jdbcTemplate.queryForObject(
-                 sql, new Object[] { user.email, user.password }, Integer.class);
+                 sql, new Object[] { user.getEmail(), user.getPassword() }, Integer.class);
 				 
 				 return id;
 		 
@@ -120,8 +127,8 @@ public class DAOimpl {
 public int insertVoiture(Voiture car){
 		
 		String query = "insert into voiture (marque, modèle, taille, id_user) values(?, ?, ?,?)";
-			int statut = this.jdbcTemplate.update(query, new Object[] { car.marque, car.modele,
-					car.taille, car.id_user});
+			int statut = this.jdbcTemplate.update(query, new Object[] { car.getMarque(), car.getModele(),
+					car.getTaille(), car.getId_user()});
 				
 			return statut;
 		
@@ -129,7 +136,7 @@ public int insertVoiture(Voiture car){
 	
 public void insertcom(Commentaires com){
 	String query ="insert into commentaires (content, note, id_place, id_user, id_writter) values (?,?,?,?,?)";
-	 this.jdbcTemplate.update(query, new Object[] {com.content, com.note, com.id_place, com.id_user, com.id_writter});
+	 this.jdbcTemplate.update(query, new Object[] {com.getContent(), com.getNote(), com.getId_place(), com.getId_user(), com.getId_writter()});
 }
