--- src/main/java/edu/isep/easypark/Reservation.java
@@ -1,42 +1,0 @@
-package edu.isep.easypark;
-
-import java.sql.Timestamp;
-
-import org.springframework.format.annotation.DateTimeFormat;
-import org.springframework.format.datetime.joda.DateTimeParser;
-
-public class Reservation {
-
-	String title;
-	Timestamp start;
-	Timestamp end;
-	
-
-	public String getTitle() {
-		return title;
-	}
-
-	public void setTitle(String title) {
-		this.title = title;
-	}
-
-	public Timestamp getStart() {
-		return start;
-	}
-
-	public void setStart(Timestamp start) {
-		this.start = start;
-	}
-
-	public Timestamp getEnd() {
-		return end;
-	}
-
-	public void setEnd(Timestamp end) {
-		this.end = end;
-	}
-
-
-
-
-}
