+++ src/main/java/edu/isep/easypark/controller/ParkingLotController.java
@@ -1,4 +1,4 @@
-package edu.isep.easypark;
+package edu.isep.easypark.controller;
 
 import java.sql.SQLException;
 import java.util.List;
@@ -17,6 +17,14 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.SessionAttributes;
 import org.springframework.web.servlet.ModelAndView;
+
+import edu.isep.easypark.DAOimpl;
+import edu.isep.easypark.ParkingLotDAOimpl;
+import edu.isep.easypark.model.Comment;
+import edu.isep.easypark.model.Commentaires;
+import edu.isep.easypark.model.ParkingLot;
+import edu.isep.easypark.model.Reservation;
+import edu.isep.easypark.model.User;
 
 @Controller
 @SessionAttributes("loggedUser")
@@ -42,11 +50,11 @@ public class ParkingLotController {
 	public String add(ParkingLot park, Model model, HttpSession session) {
 
 	
-		park.id_user=(Integer) session.getAttribute("id_user");
-		park.note=0;
+		park.setId_user((Integer) session.getAttribute("id_user"));
+		park.setNote(0);
 		
 		// ParkingLot park2 = new ParkingLot();
-		logger.info("inserting placemax " + park.ville);
+		logger.info("inserting placemax " + park.getVille());
 		int statut = dao.insertPlace(park);
 		model.addAttribute("park", park);
 		model.addAttribute("message", "Votre place à été correctement ajoutée");
@@ -88,8 +96,8 @@ public class ParkingLotController {
 	@RequestMapping(value = "/rechercher_place_advance", method = RequestMethod.POST)
 	public String rechercheAvance(Locale locale, Model model, ParkingLot searchFor) {
 
-		logger.info("Researching for this user : " + searchFor.id_user);
-		logger.info("Researching at this place : " + searchFor.adresse);
+		logger.info("Researching for this user : " + searchFor.getId_user());
+		logger.info("Researching at this place : " + searchFor.getAdresse());
 		
 		ParkingLot searchForNew = new ParkingLot();
 		model.addAttribute("searchFor", searchForNew);
@@ -120,7 +128,7 @@ public class ParkingLotController {
 		User user = (User) session.getAttribute("user");
 		
 		
-		if(user.id==parkingLot.id_user){
+		if(user.getId()==parkingLot.getId_user()){
 			
 			model.addAttribute("action", "<a href='#'>Modifier les information de votre place</a>");
 		}
@@ -130,7 +138,7 @@ public class ParkingLotController {
 		}
 		
 		
-		model.addAttribute("hasvoted",dao.hasvoted("id_place",id_place,user.id));
+		model.addAttribute("hasvoted",dao.hasvoted("id_place",id_place,user.getId()));
 //		logger.info(" "+note);
 		
 		List<Comment> list2 = dao.fillComment("id_place",id_place);
@@ -154,10 +162,18 @@ public class ParkingLotController {
 	@RequestMapping(value = "/calendar", method = RequestMethod.GET)
 	public String testCalendar(Locale locale, Model model) {
 
-		List<Reservation> list = dao2.fillReservation();
-		model.addAttribute("list", list);	
+	List<Reservation> list = dao2.fillReservation(2);
+		model.addAttribute("list", list);		
 		
 		return "testCalendar";
+	}
+	@RequestMapping(value = "/calendar2", method = RequestMethod.GET)
+	public String testCalendar2(Locale locale, Model model) {
+
+		List<Reservation> list = dao2.fillReservation(2);
+		model.addAttribute("list", list);	
+		
+		return "basic-views";
 	}
 	
 	@RequestMapping(value = "/add_commentaire2", method = RequestMethod.POST) //changer ca un de ces jouts
@@ -165,12 +181,11 @@ public class ParkingLotController {
 		
 		
 
-		logger.info("userfake "+commentaire.id_user);
-		logger.info("note "+commentaire.note);
+	
 
 		dao.insertcom(commentaire);
 
-		place(commentaire.id_place ,  model,  session);
+		place(commentaire.getId_place() ,  model,  session);
 			return "place";
 	
 	}
@@ -189,6 +204,16 @@ public class ParkingLotController {
 //	
 //	}
 	
-	
+	@RequestMapping(value = "/add_disponibilite", method = RequestMethod.GET)
+	public String addDisponibilite(@RequestParam("id_place") int id_place ,Locale locale, Model model) {
+
+
+		model.addAttribute("id_place", id_place);	
+		List<Reservation> list = dao2.fillReservation(id_place);
+		model.addAttribute("list", list);
+		
+		
+		return "add_disponibilite";
+	}
 	
 }
