+++ src/main/java/edu/isep/easypark/model/Reservation.java
@@ -1,0 +1,80 @@
+package edu.isep.easypark.model;
+
+import java.sql.Timestamp;
+
+import org.springframework.format.annotation.DateTimeFormat;
+import org.springframework.format.datetime.joda.DateTimeParser;
+
+public class Reservation {
+
+	int id;
+	String title;
+	Timestamp start;
+	Timestamp end;
+	String status;
+	int id_place;
+	int id_user;
+	
+
+	
+	
+	public int getId() {
+		return id;
+	}
+
+	public void setId(int id) {
+		this.id = id;
+	}
+
+	public String getStatus() {
+		return status;
+	}
+
+	public void setStatus(String status) {
+		this.status = status;
+	}
+
+	public int getId_place() {
+		return id_place;
+	}
+
+	public void setId_place(int id_place) {
+		this.id_place = id_place;
+	}
+
+	public int getId_user() {
+		return id_user;
+	}
+
+	public void setId_user(int id_user) {
+		this.id_user = id_user;
+	}
+
+	public String getTitle() {
+		return title;
+	}
+
+	public void setTitle(String title) {
+		this.title = title;
+	}
+
+	public Timestamp getStart() {
+		return start;
+	}
+
+	public void setStart(Timestamp start) {
+		this.start = start;
+	}
+
+	public Timestamp getEnd() {
+		return end;
+	}
+
+	public void setEnd(Timestamp end) {
+		this.end = end;
+	}
+
+
+
+
+}
