+++ src/main/java/edu/isep/easypark/model/UserInformations.java
@@ -1,0 +1,31 @@
+package edu.isep.easypark.model;
+
+public class UserInformations {
+
+	int id_user;
+	String name;
+	String firstname;
+	
+	
+	public int getId_user() {
+		return id_user;
+	}
+	public void setId_user(int id_user) {
+		this.id_user = id_user;
+	}
+	public String getName() {
+		return name;
+	}
+	public void setName(String name) {
+		this.name = name;
+	}
+	public String getFirstname() {
+		return firstname;
+	}
+	public void setFirstname(String firstname) {
+		this.firstname = firstname;
+	}
+
+	
+	
+}
