--- src/main/webapp/WEB-INF/views/ajouter_voiture.html
@@ -1,129 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-<title>Inscription</title>
-<meta charset="utf-8">
-<meta name="viewport" content="width=device-width, initial-scale=1">
-<link rel="stylesheet"
-	href="http://maxcdn.bootstrapcdn.com/bootstrap/3.2.0/css/bootstrap.min.css">
-<script
-	src="https://ajax.googleapis.com/ajax/libs/jquery/1.11.1/jquery.min.js"></script>
-<script
-	src="http://maxcdn.bootstrapcdn.com/bootstrap/3.2.0/js/bootstrap.min.js"></script>
-	<link rel="stylesheet" type="text/css" href="/resources/ajouter_voiture.css">	as
-	
-</head>
-<body>
-
-	<!-- Navigation -->
-	<nav class="navbar navbar-inverse navbar-fixed-top ">
-		<div class="container-fluid" >
-
-			<div class="navbar-header page-scroll">
-				<button type="button" class="navbar-toggle" data-toggle="collapse"data-target="#myNavbar">
-					<span class="icon-bar"></span>
-					<span class="icon-bar"></span>
-					<span class="icon-bar"></span>
-				</button>
-				
-				<a class="navbar-brand page-scroll" href="#">Easy Park</a>
-			</div>
-
-			<div class="collapse navbar-collapse" id="myNavbar">
-				<ul class="nav navbar-nav navbar-right"></ul>
-
-				<ul class="nav navbar-nav navbar-right">
-					<li class="dropdown"><a class="dropdown-toggle"data-toggle="dropdown" href="#"><span class="glyphicon glyphicon-user"></span>   Inscription</a>
-					<ul class="dropdown-menu">
-            			<li><a href="formulaire_inscription.html">En tant que particulier</a></li>
-            			<li><a href="formulaire_inscription.html">En tant que entreprise</a></li>
-          			</ul>
-				</li>
-					<li><a href="#"><span class="glyphicon glyphicon-log-in"></span>    Se connecter</a></li>
-				</ul>
-
-			</div>
-		</div>
-	</nav>
-	
-<!-- Ajouter une voiture -->
-
-<section id="ajoutervoiture">
-	<div class="container">
-
-		<h2>Ajoutez votre véhicule</h2>
-
-		<form role="form">
-
-			<div class="form-group">
-				<label for="marque">Marque:</label> <input type="text"
-					class="form-control" id="marque"
-					placeholder="Entrer la marque de votre véhicule">
-			</div>
-
-			<div class="form-group">
-				<label for="modele">Modèle:</label> <input type="text"
-					class="form-control" id="modele"
-					placeholder="Entrer le modèle de votre véhicule">
-			</div>
-
-			<div class="form-group">
-				<label for="size">Taille du véhicule:</label> <input type="text"
-					class="form-control" id="size"
-					placeholder="Entrer la taille de votre véhicule">
-			</div>
-
-			<div class="form-group">
-				<label for="color">Couleur:</label> <input type="text"
-					class="form-control" id="color"
-					placeholder="Entrer la couleur de votre véhicule">
-			</div>
-
-			<div class="form-group">
-				<label for="cgrise">La carte grise</label> <input type="file"
-					id="cgrise">
-			</div>
-
-			<div class="form-group">
-				<label for="photov">Photographie de votre véhicule</label> <input
-					type="file" id="photov">
-			</div>
-
-			<div>
-				<ul class="pagination">
-					<li><a href="#">1</a></li>
-					<li class="active"><a href="#">2</a></li>
-				</ul>
-			</div>
-
-			<div>
-				<button type="submit" class="btn btn-info btn-md">Valider</button>
-				<button type="submit" class="btn btn-info btn-md">Retourner à
-					l'étape précédente</button>
-			</div>
-			
-			<div id="test111">
-			<a href="http://www.easypark_inscription.com/">Vous avez
-					également un parking ? Cliquez ici </a> <br>
-			</div>
-
-		</form>
-	</div>
-	</section>
-	
-	<!-- Footer -->
-    <footer>
-    <nav class="navbar navbar-inverse" id="footeressai">
-        <div class="container">
-            <div class="row">
-               
-                 <div class="col-md-12">
-                    <span class="copyright">Copyright &copy; 2015</span>
-               
-                </div>   
-            </div>
-        </div>
-    </footer>
-
-</body>
-</html>
