+++ src/main/webapp/WEB-INF/views/ajouter_voiture.jsp
@@ -1,9 +1,9 @@
 <%@ page language="java" contentType="text/html; charset=ISO-8859-1"
     pageEncoding="ISO-8859-1"%>
 <!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
-<html>
+<html lang="fr">
 <head>
-<title>Ajout voiture</title>
+<title>Ajouter une voiture</title>
 <meta charset="utf-8">
 <meta name="viewport" content="width=device-width, initial-scale=1">
 <link rel="stylesheet"
@@ -12,7 +12,7 @@
 	src="https://ajax.googleapis.com/ajax/libs/jquery/1.11.1/jquery.min.js"></script>
 <script
 	src="http://maxcdn.bootstrapcdn.com/bootstrap/3.2.0/js/bootstrap.min.js"></script>
-	<link rel="stylesheet" type="text/css" href="/resources/ajouter_voiture.css">	
+	<link href="<c:url value="/resources/ajouter_voiture1.css" />" rel="stylesheet">
 	
 </head>
 <body>
