+++ src/main/webapp/WEB-INF/views/page_utilisateur.jsp
@@ -4,10 +4,24 @@
 <!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
 <html>
 <head>
-<meta http-equiv="Content-Type" content="text/html; charset=ISO-8859-1">
-<link href="<c:url value="/resources/style.css" />" rel="stylesheet"
-	type="text/css" media="screen" />
-<title>Page d'un utilisateur</title>
+	<meta http-equiv="Content-Type" content="text/html; charset=ISO-8859-1">
+	<meta name="viewport" content="width=device-width, initial-scale=1">
+	<link rel="stylesheet"
+	href="http://maxcdn.bootstrapcdn.com/bootstrap/3.2.0/css/bootstrap.min.css">
+	<script
+	src="https://ajax.googleapis.com/ajax/libs/jquery/1.11.1/jquery.min.js"></script>
+	<script
+	src="http://maxcdn.bootstrapcdn.com/bootstrap/3.2.0/js/bootstrap.min.js"></script>
+	
+
+	<link href="<c:url value="/resources/mon_profil_places1.css" />" rel="stylesheet">
+
+	<meta http-equiv="Content-Type" content="text/html; charset=ISO-8859-1">
+	
+	<link href="<c:url value="/resources/style1.css" />" rel="stylesheet">
+	
+	<title>Page d'un utilisateur</title>
+
 </head>
 <body>
 <h1>Profil de :</h1>
